youtube_tranding/requirements.txt
mysql-connector-python-rf==2.2.2
selenium>=3.141.0